requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
billiard==4.2.4
boto3==1.35.0
botocore==1.35.99
celery==5.4.0
certifi==2025.11.12
channels==4.3.2
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
Django==6.0
django-anymail==13.1
django-appconf==1.2.0
django-celery-beat==2.6.0
django-cors-headers==4.9.0
django-dbbackup==4.1.0
django-debug-toolbar==6.1.0
django-environ==0.12.0
django-filter==25.2
django-redis==6.0.0
django-storages==1.14
django-timezone-field==7.2.1
django_celery_results==2.6.0
django_compressor==4.6.0
djangorestframework==3.15.0
djangorestframework_simplejwt==5.5.1
flet==0.28.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jmespath==1.0.1
kombu==5.6.1
oauthlib==3.3.1
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-json-logger==2.0.0
pytz==2025.2
rcssmin==1.2.2
redis==5.0.4
repath==0.9.0
requests==2.32.5
rjsmin==1.2.5
s3transfer==0.10.4
sentry-sdk==2.3.0
six==1.17.0
sqlparse==0.5.4
stripe==14.0.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
uvicorn==0.38.0
vine==5.1.0
wcwidth==0.2.14
